Windows Fixes

Closes #156

See merge request f3d/f3d!155

+++ src/vtkF3DRenderPass.cxx
@@ -194,13 +194,13 @@ void vtkF3DRenderPass::Render(const vtkRenderState* s)
   this->Initialize(s);
 
   vtkRenderState backgroundState(s->GetRenderer());
-  backgroundState.SetPropArrayAndCount(this->BackgroundProps.data(), this->BackgroundProps.size());
+  backgroundState.SetPropArrayAndCount(this->BackgroundProps.data(), static_cast<int>(this->BackgroundProps.size()));
   backgroundState.SetFrameBuffer(s->GetFrameBuffer());
 
   this->BackgroundPass->Render(&backgroundState);
 
   vtkRenderState mainState(s->GetRenderer());
-  mainState.SetPropArrayAndCount(this->MainProps.data(), this->MainProps.size());
+  mainState.SetPropArrayAndCount(this->MainProps.data(), static_cast<int>(this->MainProps.size()));
   mainState.SetFrameBuffer(s->GetFrameBuffer());
 
   this->MainPass->Render(&mainState);
